backend/requirements.txt
pandoc
channels_redis
psycopg2-binary == 2.9.9
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.15.0
beautifulsoup4==4.12.3
bibtexparser==1.4.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dataclasses-json==0.6.6
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.8
django-allauth==0.63.3
django-appconf==1.0.6
django-cors-headers==4.3.1
django-eventstream==5.1.0
django-grip==3.5.0
django-imagekit==5.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docstring_parser==0.16
docutils==0.20.1
executing==2.0.1
fake-useragent==1.5.1
Flask==3.0.3
free_proxy==1.1.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.1
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
gripcontrol==4.2.0
groq==0.8.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7 
imagesize==1.4.1
incremental==22.10.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.28
langchain-google-genai==1.0.5
langchain-google-vertexai==1.0.8
langchain-groq==0.1.5
langchain-text-splitters==0.2.0
langsmith==0.1.98
latex2sympy2==1.9.1 
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pilkit==3.0
pillow==10.4.0
platformdirs==4.2.2
priority==1.3.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
pubcontrol==3.5.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1 
PyYAML==6.0.1
pyzmq==26.1.0
requests==2.32.2
requests-toolbelt==0.10.1
rsa==4.9
scholarly==1.7.11
selenium==4.22.0
service-identity==24.1.0
setuptools==70.0.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.13.2
tenacity==8.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0 
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.4.post2
